refactor(client): Move all settings clients test to same file
